+++ src/main/java/org/findy/findy_be/place/application/find/FindPlace.java
@@ -3,7 +3,7 @@ package org.findy.findy_be.place.application.find;
 import java.util.Optional;
 
 import org.findy.findy_be.place.domain.Place;
-import org.findy.findy_be.place.dto.PlaceRequest;
+import org.findy.findy_be.place.dto.request.PlaceRequest;
 
 public interface FindPlace {
