+++ src/main/java/org/findy/findy_be/place/application/find/FindPlaceService.java
@@ -3,7 +3,7 @@ package org.findy.findy_be.place.application.find;
 import java.util.Optional;
 
 import org.findy.findy_be.place.domain.Place;
-import org.findy.findy_be.place.dto.PlaceRequest;
+import org.findy.findy_be.place.dto.request.PlaceRequest;
 import org.findy.findy_be.place.repository.PlaceRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
