+++ src/main/java/org/findy/findy_be/place/application/register/RegisterPlace.java
@@ -1,6 +1,6 @@
 package org.findy.findy_be.place.application.register;
 
-import org.findy.findy_be.place.dto.PlaceRequest;
+import org.findy.findy_be.place.dto.request.PlaceRequest;
 
 public interface RegisterPlace {
