+++ src/main/java/org/findy/findy_be/place/application/register/RegisterPlaceService.java
@@ -8,7 +8,7 @@ import org.findy.findy_be.bookmark.domain.BookmarkType;
 import org.findy.findy_be.marker.application.create.CreateMarker;
 import org.findy.findy_be.place.application.find.FindPlace;
 import org.findy.findy_be.place.domain.Place;
-import org.findy.findy_be.place.dto.PlaceRequest;
+import org.findy.findy_be.place.dto.request.PlaceRequest;
 import org.findy.findy_be.place.repository.PlaceRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
